--Created a header file, fruitMap.h, to neatly house the map<> for the fruits of NORTH AMERICA. This change required a change in the return "return fruits[region]; to "return fruit.at(region);   The original logic is still present for the map used in the SOUTH AMERICA if statement, for comparison.
This begs the question; would it be easier to simply put all of the key and mapped values into a single, aggregate map<> and pull from there?

This would require a change in the findFruit() function to mantain a logical execution. A user could simply enter any location as the key and recieve the correct mapped data - so would it then be better to create one large map<> or separate map<> containers for each region.
I doubt having every location and it's corresponding fruits in one giant map<> would cause a performance issue.